+++ apps/playground/src/app/components/modal/modal-demo.component.ts
@@ -1,10 +1,8 @@
 import { Component } from '@angular/core';
-import { SkyModalService } from '@skyux/modals';
 
 @Component({
   selector: 'app-test-cmp-modal',
   templateUrl: './modal-demo.component.html',
-  providers: [SkyModalService],
 })
 export class ModalDemoComponent {
   public showHelp = false;