+++ libs/components/core/src/lib/modules/dock/dock.module.ts
@@ -4,10 +4,11 @@ import { NgModule } from '@angular/core';
 import { SkyMutationObserverService } from '../mutation/mutation-observer-service';
 
 import { SkyDockComponent } from './dock.component';
+import { SkyDockService } from './dock.service';
 
 @NgModule({
   imports: [CommonModule],
   declarations: [SkyDockComponent],
-  providers: [SkyMutationObserverService],
+  providers: [SkyDockService, SkyMutationObserverService],
 })
 export class SkyDockModule {}