+++ libs/components/lookup/src/lib/modules/selection-modal/selection-modal.module.ts
@@ -13,6 +13,7 @@ import { SkyLookupResourcesModule } from '../shared/sky-lookup-resources.module'
 
 import { SkySelectionModalItemSelectedPipe } from './selection-modal-item-selected.pipe';
 import { SkySelectionModalComponent } from './selection-modal.component';
+import { SkySelectionModalService } from './selection-modal.service';
 
 @NgModule({
   declarations: [SkySelectionModalComponent, SkySelectionModalItemSelectedPipe],
@@ -30,5 +31,6 @@ import { SkySelectionModalComponent } from './selection-modal.component';
     SkyViewkeeperModule,
     SkyWaitModule,
   ],
+  providers: [SkySelectionModalService],
 })
 export class SkySelectionModalModule {}