+++ libs/components/pages/src/lib/modules/page/page-content.component.ts
@@ -1,4 +1,15 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  ElementRef,
+  OnDestroy,
+  OnInit,
+  inject,
+} from '@angular/core';
+import {
+  SkyMediaQueryService,
+  SkyResizeObserverMediaQueryService,
+} from '@skyux/core';
 
 /**
  * Displays page contents using spacing that corresponds to the parent
@@ -9,5 +20,25 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
   template: `<ng-content />`,
   styleUrls: ['./page-content.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
+  providers: [
+    SkyResizeObserverMediaQueryService,
+    {
+      provide: SkyMediaQueryService,
+      useExisting: SkyResizeObserverMediaQueryService,
+    },
+  ],
 })
-export class SkyPageContentComponent {}
+export class SkyPageContentComponent implements OnInit, OnDestroy {
+  #elementRef = inject(ElementRef);
+  #mediaQueryService = inject(SkyResizeObserverMediaQueryService);
+
+  public ngOnInit(): void {
+    this.#mediaQueryService.observe(this.#elementRef, {
+      updateResponsiveClasses: true,
+    });
+  }
+
+  public ngOnDestroy(): void {
+    this.#mediaQueryService.unobserve();
+  }
+}